Fix imports and update tests after moving helper functions to greedybear/utils
