+++ greedybear/cronjobs/extraction/strategies/cowrie.py
@@ -8,10 +8,8 @@ from urllib.parse import urlparse
 from greedybear.consts import PAYLOAD_REQUEST, SCANNER
 from greedybear.cronjobs.extraction.strategies import BaseExtractionStrategy
 from greedybear.cronjobs.extraction.utils import (
-    get_ioc_type,
     iocs_from_hits,
     normalize_credential_field,
-    parse_timestamp,
     threatfox_submission,
 )
 from greedybear.cronjobs.repositories import (
@@ -21,6 +19,7 @@ from greedybear.cronjobs.repositories import (
 )
 from greedybear.models import IOC, CommandSequence, CowrieSession
 from greedybear.regex import REGEX_URL_PROTOCOL
+from greedybear.utils import get_ioc_type, parse_timestamp
 
 
 def parse_url_hostname(url: str) -> str | None:
