+++ greedybear/cronjobs/firehol.py
@@ -1,8 +1,8 @@
 import requests
 
 from greedybear.cronjobs.base import Cronjob
-from greedybear.cronjobs.extraction.utils import is_valid_cidr, is_valid_ipv4
 from greedybear.cronjobs.repositories import FireHolRepository
+from greedybear.utils import is_valid_cidr, is_valid_ipv4
 
 
 class FireHolCron(Cronjob):
