+++ greedybear/cronjobs/tor_exit_nodes.py
@@ -3,10 +3,10 @@ import re
 import requests
 
 from greedybear.cronjobs.base import Cronjob
-from greedybear.cronjobs.extraction.utils import is_valid_ipv4
 from greedybear.cronjobs.repositories import IocRepository
 from greedybear.cronjobs.repositories.tor import TorRepository
 from greedybear.enums import IpReputation
+from greedybear.utils import is_valid_ipv4
 
 
 class TorExitNodesCron(Cronjob):
